refactor TwoStageL3ClosClient mutex per conversation w/Raj

Approved-by: Rajagopalan Subrahmanian

+++ two_stage_l3_clos_mutex.go
@@ -13,45 +13,89 @@ const (
 )
 
 type TwoStageL3ClosMutex struct {
-	client *TwoStageL3ClosClient
-	tagId  ObjectId
+	client   *TwoStageL3ClosClient
+	tagId    ObjectId
+	readOnly bool
+	message  string
 }
 
-func (o *TwoStageL3ClosMutex) Lock(ctx context.Context, message string) error {
+// GetMessage returns the message embedded in the mutex
+func (o *TwoStageL3ClosMutex) GetMessage() string {
+	return o.message
+}
+
+// SetMessage sets the lock message embedded in the mutex
+func (o *TwoStageL3ClosMutex) SetMessage(msg string) error {
+	if o.readOnly {
+		return ApstraClientErr{
+			errType: ErrReadOnly,
+			err:     errors.New("attempt to set message of a read-only mutex"),
+		}
+	}
 	if o.tagId != "" {
-		return fmt.Errorf("attempt to lock previously locked mutex - previous lock ID '%s'", o.tagId)
+		return errors.New("attempt to set message of a locked mutex")
+	}
+	o.message = msg
+	return nil
+}
+
+// ID returns the lock ID
+func (o *TwoStageL3ClosMutex) ID() ObjectId {
+	return o.tagId
+}
+
+// Lock blocks until the mutex is locked or the context expires/is canceled.
+func (o *TwoStageL3ClosMutex) Lock(ctx context.Context) error {
+	if o.tagId != "" {
+		return ApstraClientErr{
+			errType: ErrReadOnly,
+			err:     fmt.Errorf("attempt to lock previously locked mutex - previous lock ID %q", o.tagId),
+		}
+	}
+	if o.readOnly {
+		return ApstraClientErr{
+			errType: ErrReadOnly,
+			err:     errors.New("attempt to lock read-only mutex"),
+		}
 	}
 
 	lockName := fmt.Sprintf(lockTagName, o.client.blueprintId)
 	if len(lockName) > tagNameLenMax {
-		return fmt.Errorf("lock name '%s' exceeds limit (max %d characters)", lockName, tagNameLenMax)
+		return fmt.Errorf("lock name %q exceeds limit (max %d characters)", lockName, tagNameLenMax)
 	}
 
 	// set initial LockStatus to bogus value b/c desired state is "0"
 	li := &LockInfo{LockStatus: -1}
 
-	// refuse to lock this mutex while the actual blueprint is locked
+	// refuse to lock this mutex while the apstra blueprint lock is asserted
 	var err error
 	for li.LockStatus != LockStatusUnlocked {
-		li, err = o.client.GetLockInfo(ctx)
-		if err != nil {
-			return err
-		}
-
 		// mind the timeout
 		select {
 		case <-ctx.Done():
-			return fmt.Errorf("context cancelled while waiting for lock status '%s' - %w",
+			return fmt.Errorf("context cancelled while waiting for lock status %q - %w",
 				LockStatusUnlocked.String(), ctx.Err())
 		default:
+		}
+
+		li, err = o.client.GetLockInfo(ctx)
+		if err != nil {
+			return err
 		}
 	}
 
 	// loop until we acquire the lock or the context deadline (set by caller) expires.
 	for {
+		// mind the timeout
+		select {
+		case <-ctx.Done():
+			return fmt.Errorf("context cancelled while trying to establish lock - %w", ctx.Err())
+		default:
+		}
+
 		tagId, err := o.client.client.createTag(ctx, &DesignTagRequest{
 			Label:       lockName,
-			Description: message,
+			Description: o.message,
 		})
 		if err != nil {
 			var ace ApstraClientErr
@@ -66,7 +110,14 @@ func (o *TwoStageL3ClosMutex) Lock(ctx context.Context, message string) error {
 	}
 }
 
+// Unlock releases the mutex
 func (o *TwoStageL3ClosMutex) Unlock(ctx context.Context) error {
+	if o.readOnly {
+		return ApstraClientErr{
+			errType: ErrReadOnly,
+			err:     errors.New("attempt to unlock read-only mutex"),
+		}
+	}
 	err := o.client.client.deleteTag(ctx, o.tagId)
 	if err != nil {
 		return err
@@ -75,6 +126,66 @@ func (o *TwoStageL3ClosMutex) Unlock(ctx context.Context) error {
 	return nil
 }
 
-func (o *TwoStageL3ClosMutex) Id() ObjectId {
-	return o.tagId
+// TryLock attempts to lock the mutex without blocking until success. The
+// returned boolean indicates the success of the lock attempt. On success,
+// the returned values are <true, nil, nil>.
+// When the attempt to lock is blocked by a different TwoStageL3ClosMutex,
+// a read-only copy of the offender is returned as *TwoStageL3ClosMutex for
+// inspection by the caller.
+// When the attempt to lock is blocked by Apstra's user-based blueprint lock
+// the return values are <false, nil, nil>
+func (o *TwoStageL3ClosMutex) TryLock(ctx context.Context) (bool, *TwoStageL3ClosMutex, error) {
+	if o.readOnly {
+		return false, nil, ApstraClientErr{
+			errType: ErrReadOnly,
+			err:     errors.New("attempt to unlock read-only mutex"),
+		}
+	}
+
+	lockName := fmt.Sprintf(lockTagName, o.client.blueprintId)
+	if len(lockName) > tagNameLenMax {
+		return false, nil, fmt.Errorf("lock name %q exceeds limit (max %d characters)", lockName, tagNameLenMax)
+	}
+
+	if o.tagId != "" {
+		err := o.client.client.UpdateTag(ctx, o.tagId, &DesignTagRequest{
+			Label:       lockName,
+			Description: o.message,
+		})
+		if err != nil {
+			return false, nil, fmt.Errorf("error updating tag/lock %q - %w", o.tagId, err)
+		}
+		return true, nil, nil
+	}
+
+	// refuse to lock this mutex if the apstra blueprint lock exists
+	li, err := o.client.GetLockInfo(ctx)
+	if err != nil {
+		return false, nil, err
+	}
+	if li.LockStatus != LockStatusUnlocked {
+		return false, nil, nil
+	}
+
+	tagId, err := o.client.client.createTag(ctx, &DesignTagRequest{
+		Label:       lockName,
+		Description: o.message,
+	})
+	if err != nil {
+		var ace ApstraClientErr
+		if errors.As(err, &ace) && ace.errType == ErrExists {
+			blockingTag, err := o.client.client.GetTagByLabel(ctx, lockName)
+			if err != nil {
+				return false, nil, err
+			}
+			return false, &TwoStageL3ClosMutex{
+				tagId:    blockingTag.Id,
+				readOnly: true,
+				message:  blockingTag.Data.Description,
+			}, nil
+		}
+		return false, nil, err
+	}
+	o.tagId = tagId
+	return true, nil, nil
 }
